validra-cli/requirements.txt
requests
pyyaml
click